+++ pulsar/client/transport/standard.py
@@ -1,5 +1,5 @@
 """
-Pulsar HTTP Client layer based on Python Standard Library (urllib2)
+Pulsar HTTP Client layer based on Python Standard Library (urllib)
 """
 
 from __future__ import with_statement
@@ -8,26 +8,20 @@ import mmap
 import socket
 
 from os.path import getsize
-try:
-    from urllib2 import urlopen, URLError
-except ImportError:
-    from urllib.request import urlopen
-    from urllib.error import URLError
-try:
-    from urllib2 import Request
-except ImportError:
-    from urllib.request import Request
+from urllib.request import Request, urlopen
+from urllib.error import URLError
 
 from ..exceptions import PulsarClientTransportError
 
 
-class Urllib2Transport(object):
+class UrllibTransport(object):
 
     def __init__(self, timeout=None, **kwrgs):
         self.timeout = timeout
 
     def _url_open(self, request, data):
-        return urlopen(request, data, self.timeout)
+        # data is intentionally not used here (it is part of the request object), the parameter remains for tests
+        return urlopen(request, timeout=self.timeout)
 
     def execute(self, url, method=None, data=None, input_path=None, output_path=None):
         request = self.__request(url, data, method)
@@ -40,6 +34,9 @@ class Urllib2Transport(object):
                     data = mmap.mmap(input.fileno(), 0, access=mmap.ACCESS_READ)
                 else:
                     data = b""
+                # setting the data property clears content-length, so the header must be set after (if content-length is
+                # unset, urllib sets transfer-encoding to chunked, which is not supported by webob on the server side).
+                request.data = data
                 request.add_header('Content-Length', str(size))
             try:
                 response = self._url_open(request, data)