+++ test/integration_test_state.py
@@ -14,6 +14,7 @@ from pulsar.manager_endpoint_util import (
     submit_job,
 )
 from pulsar.managers.stateful import ActiveJobs
+from pulsar.client.amqp_exchange import ACK_FORCE_NOACK_KEY
 from pulsar.client.amqp_exchange_factory import get_exchange
 from pulsar.managers.util.drmaa import DrmaaSessionFactory
 
@@ -119,6 +120,66 @@ class StateIntegrationTestCase(TempDirectoryTestCase):
 
     @skip_unless_module("kombu")
     @integration_test
+    def test_async_request_of_mq_status(self):
+        test = "async_request_of_mq_status"
+        with self._setup_app_provider(test, manager_type="queued_python") as app_provider:
+            job_id = '12345'
+
+            consumer = self._status_update_consumer(test)
+            consumer.start()
+
+            with app_provider.new_app() as app:
+                manager = app.only_manager
+                job_info = {
+                    'job_id': job_id,
+                    'command_line': 'sleep 1000',
+                    'setup': True,
+                    # Invalid staging description...
+                    'remote_staging': {"setup": [{"moo": "cow"}]}
+                }
+                # TODO: redo this with submit_job coming through MQ for test consistency.
+                submit_job(manager, job_info)
+                self._request_status(test, job_id)
+
+            import time
+            time.sleep(2)
+            consumer.wait_for_messages()
+            consumer.join()
+
+            messages = consumer.messages
+            assert len(messages) == 2, len(messages)
+            assert messages[0]["status"] == "failed"
+            assert messages[1]["status"] == "failed", messages[1]
+
+    @skip_unless_module("kombu")
+    @integration_test
+    def test_async_request_of_mq_status_lost(self):
+        test = "async_request_of_mq_status_lost"
+        with self._setup_app_provider(test, manager_type="queued_python") as app_provider:
+            job_id = '12347'  # should be lost? - never existed right?
+
+            consumer = self._status_update_consumer(test)
+            consumer.start()
+
+            with app_provider.new_app() as app:
+                app.only_manager
+                # do two messages to ensure generation of status message doesn't
+                # create a job directory we don't mean to or something like that
+                self._request_status(test, job_id)
+                self._request_status(test, job_id)
+
+            import time
+            time.sleep(2)
+            consumer.wait_for_messages()
+            consumer.join()
+
+            messages = consumer.messages
+            assert len(messages) == 2, len(messages)
+            assert messages[0]["status"] == "lost", messages[0]
+            assert messages[1]["status"] == "lost", messages[1]
+
+    @skip_unless_module("kombu")
+    @integration_test
     def test_setup_failure_fires_failed_status(self):
         test = "stating_failure_fires_failed"
         with self._setup_app_provider(test, manager_type="queued_python") as app_provider:
@@ -165,6 +226,16 @@ class StateIntegrationTestCase(TempDirectoryTestCase):
         manager = "manager_%s" % test
         consumer = SimpleConsumer(queue="status_update", url=mq_url, manager=manager)
         return consumer
+
+    def _request_status(self, test, job_id):
+        mq_url = "memory://test_%s" % test
+        manager = "manager_%s" % test
+        exchange = get_exchange(mq_url, manager, {})
+        params = {
+            "job_id": job_id,
+            ACK_FORCE_NOACK_KEY: True,
+        }
+        exchange.publish("status", params)
 
 
 class SimpleConsumer(object):